\name{Sim_Costs}
\alias{Sim_Costs}
\title{
Simulation of a cost matrix for use in NHEMOtree
}
\description{
Simulation of a cost matrix for 10 variables in combination with 'Sim_Data' and to be analyszed with "NHEMOtree".
}
\usage{
Sim_Costs()
}
\author{
Swaantje Casjens
}
\seealso{
\code{\link{Sim_Data}}, \code{\link{NHEMOtree}}
}
\examples{
Sim_Costs()
}
\keyword{Non-hierarchical evolutionary multi-objective tree learner}
